Rename monitor.py to _monitor.py and update imports
